requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
pandas>=1.3.0

# YOLO and computer vision
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
PyYAML>=6.0

# MQTT communication
paho-mqtt>=1.6.0

# PDF generation
reportlab>=3.6.0

# Data processing and utilities
requests>=2.25.0
python-dateutil>=2.8.0
geopy>=2.2.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development tools
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Optional: GPU support
# torch>=1.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html 